move child_mask.hpp to core/thread/
