+++ src/core/Makefile.am
@@ -320,7 +320,6 @@ HEADERS_COMMON                             = \
     coap/coap.hpp                            \
     coap/coap_message.hpp                    \
     coap/coap_secure.hpp                     \
-    common/child_mask.hpp                    \
     common/clearable.hpp                     \
     common/code_utils.hpp                    \
     common/crc16.hpp                         \
@@ -425,6 +424,7 @@ HEADERS_COMMON                             = \
     thread/address_resolver.hpp              \
     thread/announce_begin_server.hpp         \
     thread/announce_sender.hpp               \
+    thread/child_mask.hpp                    \
     thread/child_table.hpp                   \
     thread/discover_scanner.hpp              \
     thread/dua_manager.hpp                   \
